edit update campaign

+++ src/services/campaign.js
@@ -1,4 +1,3 @@
-const { CAMPAIGN_STATUS } = require('../constants');
 const CustomError = require('../errors/CustomError');
 const code = require('../errors/code');
 const userDao = require('../daos/user');
@@ -6,9 +5,11 @@ const intentDao = require('../daos/intent');
 const serviceDao = require('../daos/service');
 const campaignDao = require('../daos/campaign');
 const botService = require('./bot');
+const serviceCampaign = require('./serviceCampaign');
 const {
   PARTICIPATION_STATUS,
   EVENT_JOIN_CAMPAIGN,
+  CAMPAIGN_STATUS,
 } = require('../constants/index');
 
 const getCampaigns = async ({ search, fields, offset, limit, sort, query }) => {
@@ -20,7 +21,6 @@ const getCampaigns = async ({ search, fields, offset, limit, sort, query }) => {
     sort,
     query,
   });
-
   return { campaigns, count };
 };
 
@@ -29,12 +29,28 @@ const getCampaign = async (campaignId) => {
   if (!campaign)
     throw new CustomError(code.BAD_REQUEST, 'Campaign is not exists');
 
-  return campaign;
+  const response = await serviceCampaign.getServiceCampaign(
+    campaignId,
+    campaign.service.url,
+  );
+  if (!response.status) {
+    throw new Error('Campaign is not exists');
+  }
+  return { ...campaign, detailCampaign: response.result };
 };
 
-const createCampaign = async (createFields) => {
-  const { action, botId } = createFields;
-  const serviceId = createFields.service;
+const createCampaign = async ({
+  name,
+  description,
+  campaignVisibility,
+  image,
+  startTime,
+  endTime,
+  serviceId,
+  action,
+  appId,
+  botId,
+}) => {
   const service = await serviceDao.findService(serviceId);
   if (!service)
     throw new CustomError(code.BAD_REQUEST, 'Service is not exists');
@@ -45,33 +61,88 @@ const createCampaign = async (createFields) => {
       'Service does not have this action',
     );
   const campaign = await campaignDao.createCampaign({
-    ...createFields,
-    status: CAMPAIGN_STATUS.WAITING,
+    name,
+    description,
+    campaignVisibility,
+    image,
+    startTime,
+    endTime,
+    service: serviceId,
+    action,
+    appId,
+    botId,
+    status: CAMPAIGN_STATUS.DRAFT,
   });
   if (botId) {
-    const { result } = await botService.getIntents(botId);
-    const intents = result.intents.map((item) => {
-      return { ...item, campaignId: campaign.id };
-    });
-    if (intents) await intentDao.createIntents(intents);
+    const response = await botService.getIntents(botId);
+    if (response.status && response.result.intents.length) {
+      const intents = response.result.intents.map((item) => {
+        return { ...item, campaignId: campaign.id };
+      });
+      await intentDao.createIntents(intents);
+    }
   }
   return campaign;
 };
 
 const updateCampaign = async (campaignId, updateFields) => {
+  const {
+    name,
+    description,
+    campaignVisibility,
+    image,
+    startTime,
+    endTime,
+    appId,
+    botId,
+  } = updateFields;
+
   const campaignExist = await campaignDao.findCampaign(campaignId);
-  const { botId } = updateFields;
   if (!campaignExist)
     throw new CustomError(code.BAD_REQUEST, 'Campaign is not exists');
-  if (botId) {
+
+  if (botId && botId !== campaignExist.botId) {
     await intentDao.deleteIntents({ campaignId });
     const { result } = await botService.getIntents(botId);
     const intents = result.intents.map((item) => {
-      return { ...item, campaignId: campaign.id };
+      return { ...item, campaignId };
     });
-    if (intents) await intentDao.createIntents(intents);
+    if (intents.length) await intentDao.createIntents(intents);
   }
-  const campaign = await campaignDao.updateCampaign(campaignId, updateFields);
+
+  const campaign = await campaignDao.updateCampaign(campaignId, {
+    name,
+    description,
+    campaignVisibility,
+    image,
+    startTime,
+    endTime,
+    appId,
+    botId,
+  });
+  return campaign;
+};
+
+const updateServiceCampaign = async (campaignId, detailCampaign) => {
+  const campaignExist = await campaignDao.findCampaign(campaignId);
+  if (!campaignExist)
+    throw new CustomError(code.BAD_REQUEST, 'Campaign is not exists');
+
+  const result = await serviceCampaign.updateServiceCampaign(
+    campaignId,
+    campaignExist.service.url,
+    detailCampaign,
+  );
+  if (!result.status) {
+    throw new Error('Update failure');
+  }
+
+  const campaign = await campaignDao.updateCampaign(campaignId, {
+    status:
+      campaignExist.status === CAMPAIGN_STATUS.DRAFT
+        ? CAMPAIGN_STATUS.WAITING
+        : campaignExist.status,
+  });
   return campaign;
 };
 
@@ -126,6 +197,7 @@ module.exports = {
   getCampaign,
   createCampaign,
   updateCampaign,
+  updateServiceCampaign,
   deleteCampaign,
   addUser,
 };
